refactor:moving modal Image to general components
