+++ catalogo-arqueologico/src/components/ModalImage.jsx
