--- catalogo-arqueologico/src/pages/ObjectDetail/index.js
@@ -1,3 +1,0 @@
-import ObjectDetail from './ObjectDetail';
-
-export { ObjectDetail }
